+++ src/processor.rs
@@ -3,11 +3,14 @@ mod mock;
 mod simple;
 mod sink;
 
+use chrono::{DateTime, Utc};
 use std::future::Future;
 
 pub use builder::*;
 pub use simple::*;
 pub use sink::*;
+
+pub type Timestamp = DateTime<Utc>;
 
 pub trait Processor: Send + 'static {
   /// Process a log line. The line is guaranteed to be non-empty and does not contain a newline character.
@@ -17,7 +20,7 @@ pub trait Processor: Send + 'static {
   ///
   /// The line may be an [EMF](https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Embedded_Metric_Format_Specification.html) line.
   /// You can use the [`is_emf`](crate::is_emf) util function to check if it is.
-  fn process(&mut self, line: String, timestamp: i64) -> impl Future<Output = bool> + Send;
+  fn process(&mut self, line: String, timestamp: Timestamp) -> impl Future<Output = bool> + Send;
 
   /// Flushes the underlying output stream, ensuring that all intermediately buffered contents reach their destination.
   fn flush(&mut self) -> impl Future<Output = ()> + Send;